Change to how IDs are read from file and provided to new items
Found that BagItem.GetNewItemID() was actually trying to read from only the item IDs that were loaded as part of the items for the currently-open bag. This would create huge issues if someone had multiple bags going and tried to save to a new bag. Instead set method to read straight from itemStore.keys() for the source of available IDs. This helps to keep everything nice and simple, while still only having to load the bag-specific items, preventing huge RAM issues.